services/django/requirements.txt
# WallClub Core - Package compartilhado
wallclub_core @ file:///../core

Django==4.2.23
djangorestframework==3.16.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
mysqlclient==2.1.1
PyMySQL==1.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
PyJWT==2.10.1
boto3==1.34.0
requests==2.31.0
pycryptodome==3.19.0
cryptography==41.0.7
openpyxl==3.1.2
reportlab==4.0.4
bcrypt==4.1.2
pytz==2023.3
firebase-admin==6.2.0
httpx[http2]==0.25.2
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
